Move JS controller from SV to Core
To be used in SA.
